requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai-whisper==20231117
python-multipart==0.0.6
